refactor whole blockly logic to separated files
