+++ blockly/src/entities/HeroTankControlledFactory.java
@@ -1,5 +1,6 @@
 package entities;
 
+import client.HeroCommands;
 import contrib.entities.EntityFactory;
 import core.Entity;
 import core.components.PlayerComponent;
@@ -8,7 +9,6 @@ import core.components.VelocityComponent;
 import core.configuration.KeyboardConfig;
 import core.utils.components.MissingComponentException;
 import java.io.IOException;
-import server.Server;
 import utils.Direction;
 
 /**
@@ -43,11 +43,11 @@ public class HeroTankControlledFactory {
     // Add rotation controls
     pc.registerCallback(
         KeyboardConfig.MOVEMENT_LEFT.value(),
-        (entity) -> Server.instance().rotateHero(Direction.LEFT),
+        (entity) -> HeroCommands.rotate(Direction.LEFT),
         false);
     pc.registerCallback(
         KeyboardConfig.MOVEMENT_RIGHT.value(),
-        (entity) -> Server.instance().rotateHero(Direction.RIGHT),
+        (entity) -> HeroCommands.rotate(Direction.RIGHT),
         false);
 
     return hero;
